Removed watch + better widget/shortcuts integration
