+++ Threaded/Data/AppInfo.swift
@@ -9,3 +9,7 @@ public enum AppInfo {
     public static let defaultServer = "mastodon.social"
     public static let website = "https://apps.lumaa.fr/app/threaded"
 }
+
+extension AppInfo {
+    static var appVersion = Bundle.main.infoDictionary?["CFBundleShortVersionString"] as? String ?? "Unknown"
+}
