--- ThreadedWatch/Assets.xcassets/AccentColor.colorset/Contents.json
@@ -1,11 +1,0 @@
-{
-  "colors" : [
-    {
-      "idiom" : "universal"
-    }
-  ],
-  "info" : {
-    "author" : "xcode",
-    "version" : 1
-  }
-}
