--- ThreadedWatch/Assets.xcassets/AppIcon.appiconset/Contents.json
@@ -1,14 +1,0 @@
-{
-  "images" : [
-    {
-      "filename" : "ThreadedApp.png",
-      "idiom" : "universal",
-      "platform" : "watchos",
-      "size" : "1024x1024"
-    }
-  ],
-  "info" : {
-    "author" : "xcode",
-    "version" : 1
-  }
-}
