--- ThreadedWatch/Localizable.xcstrings
@@ -1,54 +1,0 @@
-{
-  "sourceLanguage" : "en",
-  "strings" : {
-    "account.followers" : {
-      "localizations" : {
-        "en" : {
-          "stringUnit" : {
-            "state" : "translated",
-            "value" : "followers"
-          }
-        },
-        "fr" : {
-          "stringUnit" : {
-            "state" : "translated",
-            "value" : "followers"
-          }
-        }
-      }
-    },
-    "iphone.login" : {
-      "localizations" : {
-        "en" : {
-          "stringUnit" : {
-            "state" : "translated",
-            "value" : "Login with your iPhone"
-          }
-        },
-        "fr" : {
-          "stringUnit" : {
-            "state" : "translated",
-            "value" : "Connexion via l'iPhone"
-          }
-        }
-      }
-    },
-    "iphone.login.description" : {
-      "localizations" : {
-        "en" : {
-          "stringUnit" : {
-            "state" : "translated",
-            "value" : "Use your iPhone to see your account on your Apple Watch"
-          }
-        },
-        "fr" : {
-          "stringUnit" : {
-            "state" : "translated",
-            "value" : "Utilisez votre iPhone pour voir vos comptes sur votre Apple Watch"
-          }
-        }
-      }
-    }
-  },
-  "version" : "1.0"
-}
