--- ThreadedWatch/ThreadedWatch Watch App.entitlements
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
-<plist version="1.0">
-<dict>
-	<key>com.apple.security.application-groups</key>
-	<array>
-		<string>group.lumaa.ThreadedApp</string>
-	</array>
-</dict>
-</plist>
