+++ ThreadedWidgets/CreatePostWidget.swift
@@ -1,0 +1,14 @@
+//// Made by Lumaa
+    
+
+import SwiftUI
+
+struct CreatePostWidget: View {
+    var body: some View {
+        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
+    }
+}
+
+#Preview {
+    CreatePostWidget()
+}
